+++ src/auth/auth.module.ts
@@ -3,14 +3,15 @@ import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { JwtModule } from '@nestjs/jwt';
 import { PrismaModule } from 'src/db/prisma.module';
+import { UserModule } from 'src/user/user.module';
 @Module({
   imports: [
-    PrismaModule,
     JwtModule.register({
       global: true,
       secret: process.env.JWT_SECRET,
       // signOptions: { expiresIn: '60s' },
     }),
+    UserModule,
   ],
   controllers: [AuthController],
   providers: [AuthService],