+++ src/event/event.module.ts
@@ -5,9 +5,10 @@ import { PrismaModule } from 'src/db/prisma.module';
 import { AuthModule } from 'src/auth/auth.module';
 import { ImageService } from './image/image.service';
 import { MemberService } from './member/member.service';
+import { UserModule } from 'src/user/user.module';
 
 @Module({
-  imports: [PrismaModule, AuthModule],
+  imports: [PrismaModule, UserModule],
   controllers: [EventController],
   providers: [EventService, ImageService, MemberService],
 })