+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # websocket-feature 14/12/24
 
 - Suppression relation model Conversation & Event: erreur de conception, les conversations privées ne sont pas liées à un event
-
-# websocket-feature 23/12/24
-
 - colonne sender_id dans la table Message
+
+# auth-refacto
+
+- suppression de la table Auth